+++ geotrek/signage/tests/test_models.py
@@ -28,6 +28,10 @@ class BladeModelTest(TestCase):
         if not user:
             UserFactory(username="__internal__")
 
+    def setUp(self):
+        self.signage = SignageFactory()
+        self.blade = BladeFactory(signage=self.signage)
+
     def test_set_topology_other_error(self):
         blade = BladeFactory.create()
         infra = InfrastructureFactory.create()