+++ docs/changelog.rst
@@ -5,10 +5,11 @@ CHANGELOG
 2.99.0+dev (XXXX-XX-XX)
 -----------------------
 
-**New features**
+**Improvements**
 
 - Filter trek and outdoor site labels according to whether they are published or not (#3529)
 - Filter by multiple structures on Blades list (#3646)
+- Add a multiselect to filter the Blades by more than one manager
 
 2.99.0 (2023-07-18)
 -----------------------